refactor(utils): move election manager filenames into filenames utils
